# web/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Payment Gateways API Microservice'
  description: 'Payment Gateways API Microservice, Version v1'
  contact:
    name: 'Support Team'
    email: support@ultainfinity.com
  version: v1
servers:
  -
    url: 'http://localhost:8100/v1'
    description: 'Payment Gateways API Microservice, Version v1'
paths:
...
  /admin/payment-setting:
    get:
      tags:
        - 'Admin / Payment-Setting'
      summary: 'Display list of all payment setting'
      description: 'Display list of all payment settings'
      operationId: d26bae4c31b7d8e9067b3ad6e5628af3
      parameters:
        -
          name: limit
          in: query
          description: 'Count of orders in response'
          required: false
          schema:
            type: integer
            default: 20
        -
          name: page
          in: query
          description: 'Page of list'
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
    post:
      tags:
        - 'Admin / Payment-Setting'
      summary: 'Method to add new payment setting'
      description: 'method to add new payment setting'
      operationId: 94a48e87dcdc60b418803ca69eefe6dc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                payment_system_id:
                  description: 'payment system id'
                  type: string
                key:
                  description: 'payment setting key'
                  type: string
                value:
                  description: 'payment setting value'
                  type: string
              type: object
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
...
components:
  securitySchemes:
    default:
      type: oauth2
      description: 'Auth Scheme'
      flows:
        implicit:
          authorizationUrl: 'https://sumraid.com/oauth2'
          scopes:
            ManagerRead: 'Manager can read'
            User: 'User access'
            ManagerWrite: 'Manager can write'